csv_import_scripts/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
gevent==24.11.1
greenlet==3.1.1
grequests==0.7.0
idna==3.10
progressbar2==4.5.0
python-utils==3.9.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
zope.event==5.0
zope.interface==7.1.1
